Version 1.4
Automatic chunk definitions for visicon-object slot values when added to the visicon to avoid warnings about chunk creation when running a model. Integration of a device instance to visicon-object instances. Visicon-object that are added to the visicon are now stored in the device instead of a the general parameter *visicon-objects*. The creation of a  visicon object now requires a device object.